Refactoring eye-tracking into its own folder.
