+++ src/eyetracking/GazeDot.tsx
