+++ src/eyetracking/WebGazerManager.js
